requirements.txt
sentence_transformers
torch
pymilvus
opencv-python
numpy
google-genai
